VM 1.0.3
writeCall should be completed, and I'm currently working on 
writeFunction and writeReturn. I have code in place for them but I'm not 
sure it's even the right code or if it's just a matter of debugging. I 
don't really understand the logic behind the way the function commands 
manipulate the stack.