VM 2.0.1
The Virtual Machine translator is done and finished for now. I would like to add functionality for parsing multiple files running the program once or parsing files in a directory. But for now it passes all the tests. I did finish the setFileName method to make those future improvements eaiser. For now, cashing it in.